refactor: controller 분리 재수정 및 api 재구성
