+++ src/test/java/cart/controller/PageAcceptanceTest.java
@@ -4,26 +4,18 @@ import static io.restassured.RestAssured.get;
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import cart.dto.ProductRequest;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.test.annotation.Rollback;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class AdminControllerTest {
+class PageAcceptanceTest {
 
     @LocalServerPort
     private int port;
@@ -31,6 +23,16 @@ class AdminControllerTest {
     @BeforeEach
     void setUp() {
         RestAssured.port = port;
+    }
+
+    @DisplayName("main 페이지에 접속하면 상태 코드로 200을 반환한다.")
+    @Test
+    void main_page() {
+        ExtractableResponse<Response> result = get("/")
+            .then()
+            .extract();
+
+        assertThat(result.statusCode()).isEqualTo(HttpStatus.OK.value());
     }
 
     @DisplayName("admin 페이지에 접속하면 상태 코드로 200을 반환한다.")
